backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
Flask-Migrate==3.1.0
Flask-CORS==3.0.10
boto3==1.26.137
botocore==1.29.137
python-dotenv==0.19.2
gunicorn==20.1.0
psycopg2-binary
requests==2.28.2
Flask-Limiter 